requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-extensions==3.2.3
django-filter==23.2
django-template-tags==1.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
gunicorn==20.1.0
httpresponses==1.0.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
netaddr==0.8.0
psycopg==3.1.8
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
template==0.7.6
timedelta==2020.12.3
toml==0.10.2
typing_extensions==4.6.3
urllib3==2.0.3
